This change does a couple of things.  First it gets the Visual Studio builds working. I added the lexing files to the VStudio projects and removed the .l files from the VStudio projects.  There was a problem with use of strtoll in TGLexer.cpp and Chris suggested switching to strtol, so that's included here.
Additionally, this checkin adds minimal x64 builds to the VStudio builds.  Build issues
related to x64 in the windows specific files for DynamicLibrary.inc and Singals.inc
are worked around, but not ultimately solved.  Binaries used to be stored in

...\win32\{Debug|Release}

but are now kept in

...\win32\bin\{win32|x64}\{Debug|Release}

intermediate files will continue to be stored in the individual project directories under 
win32.  

Some names will likely change in the future to reflect that the vstudio projects
are no longer 32-bit only, but I wanted to get things up and running today so kept away
from bigger restructuring.