Fix paq8p again...
The benchmark is compressing a file to a temporary file and then decompressing
while comparing to the original. It was using a temporary archive name based on
the input file name, and this caused races whenever make -j decided to run the
llc and nat tests in parallel.

Use basename(input).basename(argv[0]) for the temporary file name, so the
tempfiles will be file1.in.paq8p.nat and file1.in.paq8p.llc while running in
parallel.

Don't print the archive file name in the output at all, and update the reference
outputs to reflect this.