Change the representation of GNU ?: expressions to use a different expression class and to bind the shared value using OpaqueValueExpr.  This fixes an unnoticed problem with deserialization of these expressions where the deserialized form would lose the vital pointer-equality trait;  or rather, it fixes it because this patch also does the right thing for deserializing OVEs.
Change OVEs to not be a "temporary object" in the sense that copy elision is
permitted.

This new representation is not totally unawkward to work with, but I think
that's really part and parcel with the semantics we're modelling here.  In
particular, it's much easier to fix things like the copy elision bug and to
make the CFG look right.

I've tried to update the analyzer to deal with this in at least some          
obvious cases, and I think we get a much better CFG out, but the printing
of OpaqueValueExprs probably needs some work.